refactor(watch): extract useWatchEngine (reducer + stamped logging dispatch)
