+++ src/renderer/shared/hooks/watch/index.ts
@@ -10,5 +10,6 @@ export * from "./useWatchingController";
 export * from "./useWatchingSince";
 export * from "./useWatchPing";
 export * from "./watchEngine";
+export * from "./useWatchEngine";
 export * from "./useWatchSessionMeta";
 export * from "./watchStallRecovery";
